backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
langchain
langchain-openai
chromadb
pypdf
pytesseract==0.3.10
pillow==10.2.0
pandas
numpy
google-generativeai==0.3.2
groq==0.4.2
qdrant-client==1.7.0
PyPDF2==3.0.1
langchain-community==0.0.16
sentence-transformers==2.5.1
